One class for embedding evaluations
